+++ diaspy/models.py
@@ -425,6 +425,11 @@ class Comment():
 		"""
 		return self._data['author'][key]
 
+	def authordata(self):
+		"""Returns all author data of the comment.
+		"""
+		return self._data['author']
+
 class Comments():
 	def __init__(self, comments=[]):
 		self._comments = comments
@@ -492,16 +497,11 @@ class Post():
 		self.guid = guid
 		self._data = {}
 		self.comments = Comments()
-		if post_data:
-			self._data = post_data
-
-		if fetch: self._fetchdata()
+		if post_data: self._setdata(post_data)
+		elif fetch: self._fetchdata()
 		if comments:
 			if not self._data: self._fetchdata()
 			self._fetchcomments()
-		else:
-			if not self._data: self._fetchdata()
-			self.comments.set_json( self.data()['interactions']['comments'] )
 
 	def __repr__(self):
 		"""Returns string containing more information then str().
@@ -513,6 +513,11 @@ class Post():
 		"""
 		return self._data['text']
 
+	def _setdata(self, data):
+		self._data = data
+		if not bool(self.comments) and data['interactions'].get('comments', []):
+			self.comments.set_json(data['interactions'].get('comments', []))
+
 	def _fetchdata(self):
 		"""This function retrieves data of the post.
 
@@ -523,8 +528,7 @@ class Post():
 		request = self._connection.get('posts/{0}.json'.format(id))
 		if request.status_code != 200:
 			raise errors.PostError('{0}: could not fetch data for post: {1}'.format(request.status_code, id))
-		elif request:
-			self._data = request.json()
+		elif request: self._setdata(request.json());
 		return self.data()['guid']
 
 	def _fetchcomments(self):
@@ -540,12 +544,33 @@ class Post():
 			else:
 				self.comments.set([Comment(c) for c in request.json()])
 
+	def _fetchlikes(self):
+		id = self.data()['id']
+		request = self._connection.get('posts/{0}/likes.json'.format(id))
+		if request.status_code != 200:
+			raise errors.PostError('{0}: could not fetch likes for post: {1}'.format(request.status_code, id))
+		json = request.json();
+		if json: self._data['interactions']['likes'] = request.json();
+		return self._data['interactions']['likes'];
+
+	def _fetchreshares(self):
+		id = self.data()['id']
+		request = self._connection.get('posts/{0}/reshares.json'.format(id))
+		if request.status_code != 200:
+			raise errors.PostError('{0}: could not fetch likes for post: {1}'.format(request.status_code, id))
+
+		json = request.json();
+		if json: self._data['interactions']['reshares'] = request.json();
+		return self._data['interactions']['reshares'];
+
+	def fetchlikes(self): return self._fetchlikes();
+	def fetchreshares(self): return self._fetchreshares();
+
 	def fetch(self, comments = False):
 		"""Fetches post data.
 		"""
 		self._fetchdata()
-		if comments:
-			self._fetchcomments()
+		if comments: self._fetchcomments()
 		return self
 
 	def data(self, data = None):
@@ -561,7 +586,7 @@ class Post():
 		"""
 		data = {'authenticity_token': repr(self._connection)}
 
-		request = self._connection.post('posts/{0}/likes'.format(self.id),    
+		request = self._connection.post('posts/{0}/likes'.format(self.id),	
 										data=data,
 										headers={'accept': 'application/json'})
 
@@ -571,7 +596,8 @@ class Post():
 
 		likes_json = request.json()
 		if likes_json:
-			self._data['interactions']['likes'] = [likes_json]
+			self._data['interactions']['likes'].insert(0, likes_json)
+			self._data['interactions']['likes_count'] = str(int(self._data['interactions']['likes_count'])+1)
 		return likes_json
 
 	def reshare(self):
@@ -585,6 +611,11 @@ class Post():
 										headers={'accept': 'application/json'})
 		if request.status_code != 201:
 			raise Exception('{0}: Post could not be reshared'.format(request.status_code))
+
+		reshares_json = request.json()
+		if reshares_json:
+			self._data['interactions']['reshares'].insert(0, reshares_json)
+			self._data['interactions']['reshares_count'] = str(int(self._data['interactions']['reshares_count'])+1)
 		return request.json()
 
 	def comment(self, text):
@@ -625,9 +656,9 @@ class Post():
 
 	def hide(self):
 		"""
-		->    PUT /share_visibilities/42 HTTP/1.1
+		->	PUT /share_visibilities/42 HTTP/1.1
 			  post_id=123
-		<-    HTTP/1.1 200 OK
+		<-	HTTP/1.1 200 OK
 		"""
 		headers = {'x-csrf-token': repr(self._connection)}
 		params = {'post_id': json.dumps(self.id)}
@@ -638,9 +669,9 @@ class Post():
 
 	def mute(self):
 		"""
-		->    POST /blocks HTTP/1.1
+		->	POST /blocks HTTP/1.1
 			{"block":{"person_id":123}}
-		<-    HTTP/1.1 204 No Content 
+		<-	HTTP/1.1 204 No Content 
 		"""
 		headers = {'content-type':'application/json', 'x-csrf-token': repr(self._connection)}
 		data = json.dumps({ 'block': { 'person_id' : self._data['author']['id'] } })
@@ -651,8 +682,8 @@ class Post():
 
 	def subscribe(self):
 		"""
-		->    POST /posts/123/participation HTTP/1.1
-		<-    HTTP/1.1 201 Created
+		->	POST /posts/123/participation HTTP/1.1
+		<-	HTTP/1.1 201 Created
 		"""
 		headers = {'x-csrf-token': repr(self._connection)}
 		data = {}
@@ -664,9 +695,9 @@ class Post():
 
 	def unsubscribe(self):
 		"""
-		->    POST /posts/123/participation HTTP/1.1
+		->	POST /posts/123/participation HTTP/1.1
 			  _method=delete
-		<-    HTTP/1.1 200 OK
+		<-	HTTP/1.1 200 OK
 		"""
 		headers = {'x-csrf-token': repr(self._connection)}
 		data = { "_method": "delete" }
@@ -718,8 +749,32 @@ class Post():
 			raise errors.PostError('{0}: Like could not be removed.'
 								   .format(request.status_code))
 
+		self._data['interactions']['likes'].pop(0);
+		self._data['interactions']['likes_count'] = str(int(self._data['interactions']['likes_count'])-1)
+
 	def author(self, key='name'):
 		"""Returns author of the post.
 		:param key: all keys available in data['author']
 		"""
 		return self._data['author'][key]
+
+class FollowedTag():
+	"""This class represents a followed tag.
+	`diaspy.tagFollowings.TagFollowings()` uses it.
+	"""
+	def __init__(self, connection, id, name, taggings_count):
+		self._connection = connection
+		self._id, self._name, self._taggings_count = id, name, taggings_count
+
+	def id(self): return self._id
+	def name(self): return self._name
+	def count(self): return self._taggings_count
+
+	def delete(self):
+		data = {'authenticity_token': repr(self._connection)}
+		request = self._connection.delete('tag_followings/{0}'.format(self._id),
+										data=data,
+										headers={'accept': 'application/json'})
+		if request.status_code != 204:
+			raise errors.TagError('{0}: Tag could not be deleted.'
+								   .format(request.status_code))